requirements.txt
autopep8==1.5.4
click==7.1.2
dnspython==1.16.0
eventlet==0.29.1
Flask==1.1.2
Flask-PyMongo==2.3.0
Flask-SocketIO==4.3.1
greenlet==0.4.17
gunicorn==20.0.4
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
Pillow==8.0.1
pycodestyle==2.6.0
pymongo==3.11.1
python-engineio==3.13.2
python-socketio==4.6.0
scipy==1.5.4
six==1.15.0
toml==0.10.2
tqdm==4.52.0
waitress==1.4.4
Werkzeug==1.0.1
